dms任务支持根据搜索的【表名】过滤列表页面

+++ dms/admin.py
@@ -13,19 +13,16 @@ from dms.views import refresh_endpoints, refresh_tasks
 
 @admin.register(Task)
 class TaskAdmin(CommonAdmin):
-    search_fields = ('name', 'source_endpoint_arn',)
+    actions = ()
+    search_fields = ('name', 'table_name', 'source_endpoint_arn',)
     list_display = ('name', 'html_actions',)
     list_display_links = ('name',)
     readonly_fields = ('format_table_mappings',)
+    list_filter = ('table_name',)
+
+    sync_schema_task_pattern = re.compile(r'-sync-schema-source-to-redshift-onlyonce$')
 
     def has_change_permission(self, request, obj=None):
-        return False
-
-    def has_delete_permission(self, request, obj=None):
-        if obj:
-            sync_schema_task_pattern = re.compile(r'-sync-schema-source-to-redshift-onlyonce$')
-            if sync_schema_task_pattern.search(obj.name):
-                return True
         return False
 
     def get_fields(self, request, obj=None):
@@ -42,8 +39,8 @@ class TaskAdmin(CommonAdmin):
             f'<a href="{reverse("dms:refresh_tables", args=(obj.id,))}">同步的表</a>',
             f'<a href="{obj.url}">AWS控制台</a>',
         ]
-        sync_schema_task_pattern = re.compile(r'-sync-schema-source-to-redshift-onlyonce$')
-        if sync_schema_task_pattern.search(obj.name):
+
+        if self.sync_schema_task_pattern.search(obj.name):
             url = reverse(f'admin:{obj._meta.app_label}_{obj._meta.model_name}_delete', args=(obj.id,))
             buttons.append(f'<a href="{url}">删除</a>')
 
@@ -59,32 +56,14 @@ class TaskAdmin(CommonAdmin):
 
     def delete_view(self, request, object_id, extra_context=None):
         if request.method == 'POST':
-            client = boto3.client('dms')
             task = Task.objects.get(id=object_id)
-            client.delete_replication_task(ReplicationTaskArn=task.arn)
-            waiter = client.get_waiter('replication_task_deleted')
-            waiter.wait(
-                Filters=[
-                    {
-                        'Name': 'replication-task-id',
-                        'Values': [
-                            task.name
-                        ]
-                    }
-                ]
-            )
-            client.delete_endpoint(EndpointArn=task.source_endpoint_arn)
-            waiter = client.get_waiter('endpoint_deleted')
-            waiter.wait(
-                Filters=[
-                    {
-                        'Name': 'endpoint-arn',
-                        'Values': [
-                            task.source_endpoint_arn
-                        ]
-                    }
-                ]
-            )
+            if self.sync_schema_task_pattern.search(task.name):
+                client = boto3.client('dms')
+                client.delete_replication_task(ReplicationTaskArn=task.arn)
+                waiter = client.get_waiter('replication_task_deleted')
+                waiter.wait(Filters=[{'Name': 'replication-task-id', 'Values': [task.name]}])
+
+                client.delete_endpoint(EndpointArn=task.source_endpoint_arn)
         return super().delete_view(request, object_id, extra_context)
